updated edit, front end changes
i need to commit more..

+++ pal/backend/routes/events.mjs
@@ -81,12 +81,40 @@ router.post("/upload", async (req, res) => {
   let newDoc = { ...req.body, imageURL: imgurData.data.link };
   delete newDoc.path;
   const result = collection.insertOne(newDoc);
-  fs.unlink(filePath);
+
   res.status(204).send(result);
 });
 // Update the post with a new comment
 router.patch("/edit/:id", async (req, res) => {
-  console.log("edit");
+  console.log(req.body);
+
+  if (req.body.hasOwnProperty("imageURL")) {
+    const imgurForm = new FormData();
+    console.log(req.body);
+    const filePath = req.body.imageURL;
+    try {
+      imgurForm.append(
+        "image",
+        fs.readFileSync(filePath, { encoding: "base64" })
+      );
+      imgurForm.append("type", "base64");
+    } catch (e) {
+      return console.log(e);
+    }
+
+    console.log("uploading now!");
+    const imgur = await fetch("https://api.imgur.com/3/image/", {
+      method: "POST",
+      body: imgurForm,
+      headers: {
+        Authorization: `Bearer ${process.env.IMGUR_TOKEN}`,
+        Accept: "application/json",
+      },
+    });
+    const imgurData = await imgur.json();
+    fs.unlink(filePath);
+    req.body.imageURL = imgurData.data.link;
+  }
   const query = { _id: new ObjectId(req.params.id) };
   const updates = {
     $set: req.body,
@@ -94,7 +122,6 @@ router.patch("/edit/:id", async (req, res) => {
 
   let collection = await db.collection("events");
   let result = await collection.updateOne(query, updates);
-
   res.send(result).status(200);
 });
